Rename LexiconXMLView
