+++ src/lib/components/LexiconEntryView.svelte
